Extract scalar I/O helpers into yaml/scalar.py
Fourth PR in the helpers/yaml split arc (follows #796 + #798 +
#800). Moves the scalar I/O surface byte-for-byte from
yaml/__init__.py into the new yaml/scalar.py submodule:

* ESPHOME_YAML_INDENT constant
* YamlUpsertNotSupportedError exception
* _MAPPING_KEY_LINE regex + _LIST_FRAME sentinel
* _split_value_and_comment quote-aware splitter
* rewrite_yaml_scalar walker
* read_yaml_scalar (thin wrapper)
* _PLAIN_SCALAR_INDICATOR_LEAD / _FORBIDDEN_SUBSTR /
  _RESERVED_PLAIN constants
* _safe_yaml_scalar (plain-vs-quoted gate)
* _QUOTE_ESCAPES + _quote
* _strip_yaml_quotes

Updates the two sibling submodules that depended on these as
package attributes to import directly from .scalar:

* substitution.py: from .scalar import _safe_yaml_scalar,
  _strip_yaml_quotes, read_yaml_scalar, rewrite_yaml_scalar
* api_encryption.py: from .scalar import _quote,
  _strip_yaml_quotes, rewrite_yaml_scalar

Required because the old from . import ... shape would
deadlock the circular load: __init__.py's bottom-block re-exports
of .scalar haven't run yet when .api_encryption (loaded first by
the same bottom block) tries to look those up as package
attributes.

Re-exports for all seven scalar symbols + the four already in
the bottom block remain in PEP 484 redundant-alias form so
external imports (tests directly import _safe_yaml_scalar and
_strip_yaml_quotes) keep working unchanged.

__init__.py: 975 -> 683 lines (-292). scalar.py: 310 lines.

All 3404 non-e2e tests pass; ruff + mypy clean.
